requirements.txt
streamlit
pandas
numpy
plotly
google-cloud-bigquery[pandas]
db_dtypes
langchain
langchain-core
langchain-together
langchain-groq
langchain-anthropic
langchain-openai
google-generativeai
langgraph
langchain-community
seaborn
matplotlib
faiss-cpu